Added LEO. Added event listener API. Added retina problem fitness function.
Numerous performance improvements in AnjiNet and ESHyperNEATTranscriberBain.
API additions that make it easier to get (HyperNEAT)Configuration and HyperNEATEvolver singletons.

+++ src/com/anji/nn/activationfunction/GaussianActivationFunction.java
@@ -52,7 +52,7 @@ public class GaussianActivationFunction implements ActivationFunction {
 	 * @see com.anji.nn.activationfunction.ActivationFunction#apply(double)
 	 */
 	public double apply(double input) {
-		return (double) Math.exp(-(input * input * SLOPE));
+		return Math.exp(-(input * input * SLOPE));
 	}
 
 	/**
